WIP on directional conveyors
I'm not sure if it's too complex to set up, but it's much more realistic.
The user has to place conveyors considering direction, much like
transport belts in Factorio. However, they can request necessary items
or outputs requested items if it has in their inventory. It is kind of a
mix of transport belts and logistic network.