An attempt to use building hash digest to update only changed ones
It is the same idea as the tile chunks digest, but it's more tricky
for buildings, because they contain HashMaps, which are not hashable.
We collect them to a Vec and sort, but still doesn't seem that the
hash is compatible between the server and wasm.
I'm not sure if it's because of architecture difference.