Add time measurement code and slightly optimize path finding for delivery
We found out that path finding logic takes the longest time, especially
after the base has expanded and the crews has a lot of tiles to search.
It can take few hundred milliseconds in the server and few seconds in
the Wasm client.

Although our main target was not the path finding in this branch, our
bottleneck was not the building lifetime management.