Use RefCell to avoid slice tricks
Now it's much more ergonomic and easy to read, but in exchange,
every EntityEntry is wrapped in a RefCell dynamic borrow checking.
It's unclear if it's worth it, but our experience so far tells
that the overhead of RefCell is quite low, when combined with a
data more than dozens of bytes.

However, there was another unexpected drawback of using RefCell.
Now that we have RefCell as a part of AsteroidColoniesGame, it
cannot be Sync (we would need Mutex to make it Sync).
It is a problem in the server where we would like to use RwLock
to wrap AsteroidColoniesGame, because RwLock requires Sync bound
for the internal type.
The workaround is either to replace the RwLock with a Mutex or
to replace all RefCells inside AsteroidColoniesGame with Mutexes.
Neither are great, but probably the former has less overhead,
since we do not have so many readers at the same time.